fix: keep dd helper compatible with unity builds
